+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Changed wide format places definition for older release versions
+- Changed get all columns function for wide format places definition
 - Bug where code crashed when release index hit zero matches [#11](https://github.com/kraina-ai/overturemaestro/issues/11)
 
 ## [0.2.2] - 2025-01-17